+++ changelog.md
@@ -19,4 +19,5 @@
 
 ## Next
 
+* Tests for creation of express routes
 * Refactor tests